Add cargo feature to control canvas backend.
