+++ components/canvas/Cargo.toml
@@ -11,10 +11,13 @@ name = "canvas"
 path = "lib.rs"
 
 [features]
+azure_backend = ["azure"]
+default = ["azure_backend"]
 webgl_backtrace = ["canvas_traits/webgl_backtrace"]
+raqote_backend = ["raqote"]
 
 [dependencies]
-azure = {git = "https://github.com/servo/rust-azure"}
+azure = {git = "https://github.com/servo/rust-azure", optional = true}
 byteorder = "1"
 canvas_traits = {path = "../canvas_traits"}
 compositing = {path = "../compositing"}
@@ -27,7 +30,7 @@ ipc-channel = "0.11"
 log = "0.4"
 num-traits = "0.2"
 offscreen_gl_context = {version = "0.22", features = ["serde", "osmesa"]}
-raqote = {git = "https://github.com/jrmuizel/raqote"}
+raqote = {git = "https://github.com/jrmuizel/raqote", optional = true}
 pixels = {path = "../pixels"}
 serde_bytes = "0.10"
 servo_config = {path = "../config"}
