+++ components/canvas/canvas_data.rs
@@ -2,18 +2,13 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-use azure::azure::{AzFloat, AzGradientStop, AzIntSize, AzPoint};
-use azure::azure_hl;
-use azure::azure_hl::SurfacePattern;
-use azure::azure_hl::{AntialiasMode, AsAzurePoint, CapStyle, JoinStyle};
-use azure::azure_hl::{BackendType, DrawTarget};
-use azure::azure_hl::{ColorPattern, Filter};
-use azure::azure_hl::{LinearGradientPattern, RadialGradientPattern};
+use crate::canvas_paint_thread::AntialiasMode;
 use canvas_traits::canvas::*;
 use cssparser::RGBA;
 use euclid::{Point2D, Rect, Size2D, Transform2D, Vector2D};
 use ipc_channel::ipc::{IpcSender, IpcSharedMemory};
 use num_traits::ToPrimitive;
+use std::marker::PhantomData;
 use std::mem;
 use std::sync::Arc;
 use webrender::api::DirtyRect;
@@ -57,9 +52,21 @@ impl PathState {
     }
 }
 
+pub trait Backend {
+    fn get_composition_op(&self, opts: &DrawOptions) -> CompositionOp;
+    fn need_to_draw_shadow(&self, color: &Color) -> bool;
+    fn set_shadow_color<'a>(&mut self, color: RGBA, state: &mut CanvasPaintState<'a>);
+    fn set_fill_style<'a>(&mut self, style: FillOrStrokeStyle, state: &mut CanvasPaintState<'a>, drawtarget: &GenericDrawTarget);
+    fn set_stroke_style<'a>(&mut self, style: FillOrStrokeStyle, state: &mut CanvasPaintState<'a>, drawtarget: &GenericDrawTarget);
+    fn set_global_composition<'a>(&mut self, op: CompositionOrBlending, state: &mut CanvasPaintState<'a>);
+    fn create_drawtarget(&self, size: Size2D<u64>) -> Box<GenericDrawTarget>;
+    fn recreate_paint_state<'a>(&self, state: &CanvasPaintState<'a>) -> CanvasPaintState<'a>;
+    fn size_from_pattern(&self, rect: &Rect<f32>, pattern: &Pattern) -> Option<Size2D<f32>>;
+}
+
 /// A generic PathBuilder that abstracts the interface for
 /// azure's and raqote's PathBuilder.
-trait GenericPathBuilder {
+pub trait GenericPathBuilder {
     fn arc(
         &self,
         origin: Point2D<f32>,
@@ -92,93 +99,20 @@ trait GenericPathBuilder {
     fn finish(&self) -> Path;
 }
 
-impl GenericPathBuilder for azure_hl::PathBuilder {
-    fn arc(
-        &self,
-        origin: Point2D<f32>,
-        radius: f32,
-        start_angle: f32,
-        end_angle: f32,
-        anticlockwise: bool,
-    ) {
-        self.arc(
-            origin as Point2D<AzFloat>,
-            radius as AzFloat,
-            start_angle as AzFloat,
-            end_angle as AzFloat,
-            anticlockwise,
-        );
-    }
-    fn bezier_curve_to(
-        &self,
-        control_point1: &Point2D<f32>,
-        control_point2: &Point2D<f32>,
-        control_point3: &Point2D<f32>,
-    ) {
-        self.bezier_curve_to(
-            control_point1 as &Point2D<AzFloat>,
-            control_point2 as &Point2D<AzFloat>,
-            control_point3 as &Point2D<AzFloat>,
-        );
-    }
-    fn close(&self) {
-        self.close();
-    }
-    fn ellipse(
-        &self,
-        origin: Point2D<f32>,
-        radius_x: f32,
-        radius_y: f32,
-        rotation_angle: f32,
-        start_angle: f32,
-        end_angle: f32,
-        anticlockwise: bool,
-    ) {
-        self.ellipse(
-            origin as Point2D<AzFloat>,
-            radius_x as AzFloat,
-            radius_y as AzFloat,
-            rotation_angle as AzFloat,
-            start_angle as AzFloat,
-            end_angle as AzFloat,
-            anticlockwise,
-        );
-    }
-    fn get_current_point(&self) -> Point2D<f32> {
-        let AzPoint { x, y } = self.get_current_point();
-        Point2D::new(x as f32, y as f32)
-    }
-    fn line_to(&self, point: Point2D<f32>) {
-        self.line_to(point as Point2D<AzFloat>);
-    }
-    fn move_to(&self, point: Point2D<f32>) {
-        self.move_to(point as Point2D<AzFloat>);
-    }
-    fn quadratic_curve_to(&self, control_point: &Point2D<f32>, end_point: &Point2D<f32>) {
-        self.quadratic_curve_to(
-            control_point as &Point2D<AzFloat>,
-            end_point as &Point2D<AzFloat>,
-        );
-    }
-    fn finish(&self) -> Path {
-        Path::Azure(self.finish())
-    }
-}
-
 /// A wrapper around a stored PathBuilder and an optional transformation that should be
 /// applied to any points to ensure they are in the matching device space.
 struct PathBuilderRef<'a> {
     builder: &'a Box<GenericPathBuilder>,
-    transform: Transform2D<AzFloat>,
+    transform: Transform2D<f32>,
 }
 
 impl<'a> PathBuilderRef<'a> {
-    fn line_to(&self, pt: &Point2D<AzFloat>) {
+    fn line_to(&self, pt: &Point2D<f32>) {
         let pt = self.transform.transform_point(pt);
         self.builder.line_to(pt);
     }
 
-    fn move_to(&self, pt: &Point2D<AzFloat>) {
+    fn move_to(&self, pt: &Point2D<f32>) {
         let pt = self.transform.transform_point(pt);
         self.builder.move_to(pt);
     }
@@ -202,7 +136,7 @@ impl<'a> PathBuilderRef<'a> {
         self.builder.close();
     }
 
-    fn quadratic_curve_to(&self, cp: &Point2D<AzFloat>, endpoint: &Point2D<AzFloat>) {
+    fn quadratic_curve_to(&self, cp: &Point2D<f32>, endpoint: &Point2D<f32>) {
         self.builder.quadratic_curve_to(
             &self.transform.transform_point(cp),
             &self.transform.transform_point(endpoint),
@@ -211,9 +145,9 @@ impl<'a> PathBuilderRef<'a> {
 
     fn bezier_curve_to(
         &self,
-        cp1: &Point2D<AzFloat>,
-        cp2: &Point2D<AzFloat>,
-        endpoint: &Point2D<AzFloat>,
+        cp1: &Point2D<f32>,
+        cp2: &Point2D<f32>,
+        endpoint: &Point2D<f32>,
     ) {
         self.builder.bezier_curve_to(
             &self.transform.transform_point(cp1),
@@ -224,10 +158,10 @@ impl<'a> PathBuilderRef<'a> {
 
     fn arc(
         &self,
-        center: &Point2D<AzFloat>,
-        radius: AzFloat,
-        start_angle: AzFloat,
-        end_angle: AzFloat,
+        center: &Point2D<f32>,
+        radius: f32,
+        start_angle: f32,
+        end_angle: f32,
         ccw: bool,
     ) {
         let center = self.transform.transform_point(center);
@@ -237,12 +171,12 @@ impl<'a> PathBuilderRef<'a> {
 
     pub fn ellipse(
         &self,
-        center: &Point2D<AzFloat>,
-        radius_x: AzFloat,
-        radius_y: AzFloat,
-        rotation_angle: AzFloat,
-        start_angle: AzFloat,
-        end_angle: AzFloat,
+        center: &Point2D<f32>,
+        radius_x: f32,
+        radius_y: f32,
+        rotation_angle: f32,
+        start_angle: f32,
+        end_angle: f32,
         ccw: bool,
     ) {
         let center = self.transform.transform_point(center);
@@ -257,7 +191,7 @@ impl<'a> PathBuilderRef<'a> {
         );
     }
 
-    fn current_point(&self) -> Option<Point2D<AzFloat>> {
+    fn current_point(&self) -> Option<Point2D<f32>> {
         let inverse = match self.transform.inverse() {
             Some(i) => i,
             None => return None,
@@ -270,7 +204,7 @@ impl<'a> PathBuilderRef<'a> {
 // TODO(pylbrecht)
 // This defines required methods for DrawTarget of azure and raqote
 // The prototypes are derived from azure's methods.
-trait GenericDrawTarget {
+pub trait GenericDrawTarget {
     fn clear_rect(&self, rect: &Rect<f32>);
     fn copy_surface(&self, surface: SourceSurface, source: Rect<i32>, destination: Point2D<i32>);
     fn create_gradient_stops(
@@ -289,15 +223,14 @@ trait GenericDrawTarget {
         data: &[u8],
         size: Size2D<i32>,
         stride: i32,
-        format: SurfaceFormat,
     ) -> Option<SourceSurface>;
     fn draw_surface(
         &self,
         surface: SourceSurface,
-        dest: Rect<f32>,
-        source: Rect<f32>,
-        surf_options: DrawSurfaceOptions,
-        options: DrawOptions,
+        dest: Rect<f64>,
+        source: Rect<f64>,
+        filter: Filter,
+        draw_options: &DrawOptions,
     );
     fn draw_surface_with_shadow(
         &self,
@@ -311,7 +244,7 @@ trait GenericDrawTarget {
     fn fill(&self, path: &Path, pattern: Pattern, draw_options: &DrawOptions);
     fn fill_rect(&self, rect: &Rect<f32>, pattern: Pattern, draw_options: Option<&DrawOptions>);
     fn get_format(&self) -> SurfaceFormat;
-    fn get_size(&self) -> IntSize;
+    fn get_size(&self) -> Size2D<i32>;
     fn get_transform(&self) -> Transform2D<f32>;
     fn pop_clip(&self);
     fn push_clip(&self, path: &Path);
@@ -339,394 +272,109 @@ trait GenericDrawTarget {
         stroke_options: &StrokeOptions,
         draw_options: &DrawOptions,
     );
-}
-
-impl GenericDrawTarget for azure_hl::DrawTarget {
-    fn clear_rect(&self, rect: &Rect<f32>) {
-        self.clear_rect(rect as &Rect<AzFloat>);
-    }
-
-    fn copy_surface(&self, surface: SourceSurface, source: Rect<i32>, destination: Point2D<i32>) {
-        self.copy_surface(surface.into_azure(), source, destination);
-    }
-
-    fn create_gradient_stops(
-        &self,
-        gradient_stops: Vec<GradientStop>,
-        extend_mode: ExtendMode,
-    ) -> GradientStops {
-        let gradient_stops: Vec<AzGradientStop> =
-            gradient_stops.into_iter().map(|x| x.into_azure()).collect();
-        GradientStops::Azure(self.create_gradient_stops(&gradient_stops, extend_mode.into_azure()))
-    }
-
-    fn create_path_builder(&self) -> Box<GenericPathBuilder> {
-        Box::new(self.create_path_builder())
-    }
-
-    fn create_similar_draw_target(
-        &self,
-        size: &Size2D<i32>,
-        format: SurfaceFormat,
-    ) -> Box<GenericDrawTarget> {
-        Box::new(self.create_similar_draw_target(size, format.into_azure()))
-    }
-    fn create_source_surface_from_data(
-        &self,
-        data: &[u8],
-        size: Size2D<i32>,
-        stride: i32,
-        format: SurfaceFormat,
-    ) -> Option<SourceSurface> {
-        self.create_source_surface_from_data(data, size, stride, format.into_azure()).map(|s| SourceSurface::Azure(s))
-    }
-    fn draw_surface(
-        &self,
-        surface: SourceSurface,
-        dest: Rect<f32>,
-        source: Rect<f32>,
-        surf_options: DrawSurfaceOptions,
-        options: DrawOptions,
-    ) {
-        self.draw_surface(
-            surface.into_azure(),
-            dest as Rect<AzFloat>,
-            source as Rect<AzFloat>,
-            surf_options.into_azure(),
-            options.into_azure(),
-        );
-    }
-    fn draw_surface_with_shadow(
-        &self,
-        surface: SourceSurface,
-        dest: &Point2D<f32>,
-        color: &Color,
-        offset: &Vector2D<f32>,
-        sigma: f32,
-        operator: CompositionOp,
-    ) {
-        self.draw_surface_with_shadow(
-            surface.into_azure(),
-            dest as &Point2D<AzFloat>,
-            color.as_azure(),
-            offset as &Vector2D<AzFloat>,
-            sigma as AzFloat,
-            operator.into_azure(),
-        );
-    }
-    fn fill(&self, path: &Path, pattern: Pattern, draw_options: &DrawOptions) {
-        self.fill(
-            path.as_azure(),
-            pattern.as_azure().to_pattern_ref(),
-            draw_options.as_azure(),
-        );
-    }
-    fn fill_rect(&self, rect: &Rect<f32>, pattern: Pattern, draw_options: Option<&DrawOptions>) {
-        self.fill_rect(
-            rect as &Rect<AzFloat>,
-            pattern.as_azure().to_pattern_ref(),
-            draw_options.map(|x| x.as_azure()),
-        );
-    }
-    fn get_format(&self) -> SurfaceFormat {
-        SurfaceFormat::Azure(self.get_format())
-    }
-    fn get_size(&self) -> IntSize {
-        IntSize::Azure(self.get_size())
-    }
-    fn get_transform(&self) -> Transform2D<f32> {
-        self.get_transform() as Transform2D<f32>
-    }
-    fn pop_clip(&self) {
-        self.pop_clip();
-    }
-    fn push_clip(&self, path: &Path) {
-        self.push_clip(path.as_azure());
-    }
-    fn set_transform(&self, matrix: &Transform2D<f32>) {
-        self.set_transform(matrix as &Transform2D<AzFloat>);
-    }
-    fn snapshot(&self) -> SourceSurface {
-        SourceSurface::Azure(self.snapshot())
-    }
-    fn stroke(
-        &self,
-        path: &Path,
-        pattern: Pattern,
-        stroke_options: &StrokeOptions,
-        draw_options: &DrawOptions,
-    ) {
-        self.stroke(
-            path.as_azure(),
-            pattern.as_azure().to_pattern_ref(),
-            stroke_options.as_azure(),
-            draw_options.as_azure(),
-        );
-    }
-    fn stroke_line(
-        &self,
-        start: Point2D<f32>,
-        end: Point2D<f32>,
-        pattern: Pattern,
-        stroke_options: &StrokeOptions,
-        draw_options: &DrawOptions,
-    ) {
-        self.stroke_line(
-            start as Point2D<AzFloat>,
-            end as Point2D<AzFloat>,
-            pattern.as_azure().to_pattern_ref(),
-            stroke_options.as_azure(),
-            draw_options.as_azure(),
-        );
-    }
-    fn stroke_rect(
-        &self,
-        rect: &Rect<f32>,
-        pattern: Pattern,
-        stroke_options: &StrokeOptions,
-        draw_options: &DrawOptions,
-    ) {
-        self.stroke_rect(
-            rect as &Rect<AzFloat>,
-            pattern.as_azure().to_pattern_ref(),
-            stroke_options.as_azure(),
-            draw_options.as_azure(),
-        );
-    }
+    fn snapshot_data(&self) -> Vec<u8>;
 }
 
 #[derive(Clone)]
-enum ExtendMode {
+pub enum ExtendMode {
+    #[cfg(feature = "azure_backend")]
     Azure(azure_hl::ExtendMode),
+    #[cfg(feature = "raqote_backend")]
     Raqote(()),
 }
 
-impl ExtendMode {
-    fn into_azure(self) -> azure_hl::ExtendMode {
-        match self {
-            ExtendMode::Azure(m) => m,
-            _ => unreachable!(),
-        }
-    }
-}
-
-enum GradientStop {
+pub enum GradientStop {
+    #[cfg(feature = "azure_backend")]
     Azure(AzGradientStop),
+    #[cfg(feature = "raqote_backend")]
     Raqote(()),
 }
 
-impl GradientStop {
-    fn into_azure(self) -> AzGradientStop {
-        match self {
-            GradientStop::Azure(s) => s,
-            _ => unreachable!(),
-        }
-    }
-}
-
-enum GradientStops {
+pub enum GradientStops {
+    #[cfg(feature = "azure_backend")]
     Azure(azure_hl::GradientStops),
+    #[cfg(feature = "raqote_backend")]
     Raqote(()),
-}
-
-impl GradientStops {
-    fn into_azure(self) -> azure_hl::GradientStops {
-        match self {
-            GradientStops::Azure(s) => s,
-            _ => unreachable!(),
-        }
-    }
 }
 
 #[derive(Clone)]
 pub enum Color {
+    #[cfg(feature = "azure_backend")]
     Azure(azure_hl::Color),
+    #[cfg(feature = "raqote_backend")]
     Raqote(()),
 }
 
-impl Color {
-    fn as_azure(&self) -> &azure_hl::Color {
-        match self {
-            Color::Azure(s) => s,
-            _ => unreachable!(),
-        }
-    }
-}
 
 #[derive(Clone)]
-enum CompositionOp {
+pub enum CompositionOp {
+    #[cfg(feature = "azure_backend")]
     Azure(azure_hl::CompositionOp),
+    #[cfg(feature = "raqote_backend")]
     Raqote(()),
 }
 
-impl CompositionOp {
-    fn into_azure(self) -> azure_hl::CompositionOp {
-        match self {
-            CompositionOp::Azure(s) => s,
-            _ => unreachable!(),
-        }
-    }
-}
-
-enum SurfaceFormat {
+pub enum SurfaceFormat {
+    #[cfg(feature = "azure_backend")]
     Azure(azure_hl::SurfaceFormat),
+    #[cfg(feature = "raqote_backend")]
     Raqote(()),
 }
 
-impl SurfaceFormat {
-    fn into_azure(self) -> azure_hl::SurfaceFormat {
-        match self {
-            SurfaceFormat::Azure(s) => s,
-            _ => unreachable!(),
-        }
-    }
-}
-
 #[derive(Clone)]
-enum SourceSurface {
+pub enum SourceSurface {
+    #[cfg(feature = "azure_backend")]
     Azure(azure_hl::SourceSurface),
+    #[cfg(feature = "raqote_backend")]
     Raqote(()),
 }
 
-impl SourceSurface {
-    fn into_azure(self) -> azure_hl::SourceSurface {
-        match self {
-            SourceSurface::Azure(s) => s,
-            _ => unreachable!(),
-        }
-    }
-}
-
-enum IntSize {
-    Azure(AzIntSize),
-    Raqote(()),
-}
-
-impl IntSize {
-    fn into_azure(self) -> AzIntSize {
-        match self {
-            IntSize::Azure(s) => s,
-            _ => unreachable!(),
-        }
-    }
-}
-
-enum Path {
+pub enum Path {
+    #[cfg(feature = "azure_backend")]
     Azure(azure_hl::Path),
+    #[cfg(feature = "raqote_backend")]
     Raqote(()),
 }
 
-impl Path {
-    fn as_azure(&self) -> &azure_hl::Path {
-        match self {
-            Path::Azure(p) => p,
-            _ => unreachable!(),
-        }
-    }
-}
-
 #[derive(Clone)]
-enum Pattern {
+pub enum Pattern {
+    #[cfg(feature = "azure_backend")]
     Azure(azure_hl::Pattern),
+    #[cfg(feature = "raqote_backend")]
     Raqote(()),
 }
 
-impl Pattern {
-    fn as_azure(&self) -> &azure_hl::Pattern {
-        match self {
-            Pattern::Azure(p) => p,
-            _ => unreachable!(),
-        }
-    }
-}
-
-enum DrawSurfaceOptions {
+pub enum DrawSurfaceOptions {
+    #[cfg(feature = "azure_backend")]
     Azure(azure_hl::DrawSurfaceOptions),
+    #[cfg(feature = "raqote_backend")]
     Raqote(()),
 }
 
-impl DrawSurfaceOptions {
-    fn into_azure(self) -> azure_hl::DrawSurfaceOptions {
-        match self {
-            DrawSurfaceOptions::Azure(options) => options,
-            _ => unreachable!(),
-        }
-    }
-}
-
 #[derive(Clone)]
-enum DrawOptions {
+pub enum DrawOptions {
+    #[cfg(feature = "azure_backend")]
     Azure(azure_hl::DrawOptions),
+    #[cfg(feature = "raqote_backend")]
     Raqote(()),
-}
-
-impl DrawOptions {
-    fn as_azure(&self) -> &azure_hl::DrawOptions {
-        match self {
-            DrawOptions::Azure(options) => options,
-            _ => unreachable!(),
-        }
-    }
-    fn as_azure_mut(&mut self) -> &mut azure_hl::DrawOptions {
-        match self {
-            DrawOptions::Azure(options) => options,
-            _ => unreachable!(),
-        }
-    }
-    fn into_azure(self) -> azure_hl::DrawOptions {
-        match self {
-            DrawOptions::Azure(options) => options,
-            _ => unreachable!(),
-        }
-    }
-    fn set_alpha(&mut self, val: f32) {
-        match self {
-            DrawOptions::Azure(options) => options.alpha = val as AzFloat,
-            _ => unreachable!(),
-        }
-    }
 }
 
 #[derive(Clone)]
-enum StrokeOptions<'a> {
+pub enum StrokeOptions<'a> {
+    #[cfg(feature = "azure_backend")]
     Azure(azure_hl::StrokeOptions<'a>),
-    Raqote(()),
+    #[cfg(feature = "raqote_backend")]
+    Raqote(PhantomData<&'a ()>),
 }
 
-impl<'a> StrokeOptions<'a> {
-    fn as_azure(&self) -> &azure_hl::StrokeOptions<'a> {
-        match self {
-            StrokeOptions::Azure(options) => options,
-            _ => unreachable!(),
-        }
-    }
-    fn set_line_width(&mut self, val: f32) {
-        match self {
-            StrokeOptions::Azure(options) => options.line_width = val as AzFloat,
-            _ => unreachable!(),
-        }
-    }
-    fn set_miter_limit(&mut self, val: f32) {
-        match self {
-            StrokeOptions::Azure(options) => options.miter_limit = val as AzFloat,
-            _ => unreachable!(),
-        }
-    }
-    fn set_line_join(&mut self, val: LineJoinStyle) {
-        match self {
-            StrokeOptions::Azure(options) => options.line_join = val.to_azure_style(),
-            _ => unreachable!(),
-        }
-    }
-    fn set_line_cap(&mut self, val: LineCapStyle) {
-        match self {
-            StrokeOptions::Azure(options) => options.line_cap = val.to_azure_style(),
-            _ => unreachable!(),
-        }
-    }
+#[derive(Clone, Copy)]
+pub enum Filter {
+    Linear,
+    Point,
 }
 
 pub struct CanvasData<'a> {
+    backend: Box<Backend>,
     drawtarget: Box<GenericDrawTarget>,
     path_state: Option<PathState>,
     state: CanvasPaintState<'a>,
@@ -740,6 +388,16 @@ pub struct CanvasData<'a> {
     pub canvas_id: CanvasId,
 }
 
+#[cfg(feature = "azure_backend")]
+fn create_backend() -> Box<Backend> {
+    Box::new(crate::azure_backend::AzureBackend)
+}
+
+#[cfg(feature = "raqote_backend")]
+fn create_backend() -> Box<Backend> {
+    Box::new(crate::raqote_backend::RaqoteBackend)
+}
+
 impl<'a> CanvasData<'a> {
     pub fn new(
         size: Size2D<u64>,
@@ -747,12 +405,14 @@ impl<'a> CanvasData<'a> {
         antialias: AntialiasMode,
         canvas_id: CanvasId,
     ) -> CanvasData<'a> {
-        let draw_target = CanvasData::create(size);
+        let backend = create_backend();
+        let draw_target = backend.create_drawtarget(size);;
         let webrender_api = webrender_api_sender.create_api();
         CanvasData {
+            backend,
             drawtarget: draw_target,
             path_state: None,
-            state: CanvasPaintState::new(antialias, CanvasBackend::Azure),
+            state: CanvasPaintState::new(antialias),
             saved_states: vec![],
             webrender_api: webrender_api,
             image_key: None,
@@ -780,7 +440,6 @@ impl<'a> CanvasData<'a> {
         };
 
 
-        // TODO(pylbrecht) create another clousure for raqote
         let writer = |draw_target: &GenericDrawTarget| {
             write_image(
                 draw_target,
@@ -788,8 +447,7 @@ impl<'a> CanvasData<'a> {
                 source_rect.size,
                 dest_rect,
                 smoothing_enabled,
-                CompositionOp::Azure(self.state.draw_options.as_azure().composition),
-                self.state.draw_options.as_azure().alpha,
+                &self.state.draw_options,
             );
         };
 
@@ -826,31 +484,15 @@ impl<'a> CanvasData<'a> {
     }
 
     pub fn fill_rect(&self, rect: &Rect<f32>) {
-        if is_zero_size_gradient(&self.state.fill_style) {
+        if self.state.fill_style.is_zero_size_gradient() {
             return; // Paint nothing if gradient size is zero.
         }
 
         let draw_rect = Rect::new(
             rect.origin,
-            match &self.state.fill_style {
-                Pattern::Azure(pattern) => {
-                    match pattern {
-                        azure_hl::Pattern::Surface(ref surface) => {
-                            let surface_size = surface.size();
-                            match (surface.repeat_x, surface.repeat_y) {
-                                (true, true) => rect.size,
-                                (true, false) => Size2D::new(rect.size.width, surface_size.height as f32),
-                                (false, true) => Size2D::new(surface_size.width as f32, rect.size.height),
-                                (false, false) => {
-                                    Size2D::new(surface_size.width as f32, surface_size.height as f32)
-                                },
-                            }
-                        }
-                        _ => rect.size,
-                    }
-                },
-                _ => unreachable!(),
-            }
+            self.backend
+                .size_from_pattern(&rect, &self.state.fill_style)
+                .unwrap_or(rect.size),
         );
 
         if self.need_to_draw_shadow() {
@@ -875,7 +517,7 @@ impl<'a> CanvasData<'a> {
     }
 
     pub fn stroke_rect(&self, rect: &Rect<f32>) {
-        if is_zero_size_gradient(&self.state.stroke_style) {
+        if self.state.stroke_style.is_zero_size_gradient() {
             return; // Paint nothing if gradient size is zero.
         }
 
@@ -889,23 +531,11 @@ impl<'a> CanvasData<'a> {
                 );
             });
         } else if rect.size.width == 0. || rect.size.height == 0. {
-            let cap = match self.state.stroke_opts.as_azure().line_join {
-                JoinStyle::Round => CapStyle::Round,
-                _ => CapStyle::Butt,
-            };
-
-            let stroke_opts = StrokeOptions::Azure(azure_hl::StrokeOptions::new(
-                self.state.stroke_opts.as_azure().line_width,
-                self.state.stroke_opts.as_azure().line_join,
-                cap,
-                self.state.stroke_opts.as_azure().miter_limit,
-                self.state.stroke_opts.as_azure().mDashPattern,
-            ));
             self.drawtarget.stroke_line(
                 rect.origin,
                 rect.bottom_right(),
                 self.state.stroke_style.clone(),
-                &stroke_opts,
+                &self.state.stroke_opts,
                 &self.state.draw_options,
             );
         } else {
@@ -951,7 +581,7 @@ impl<'a> CanvasData<'a> {
         // any transform is present.
         let new_state = match *self.path_state.as_ref().unwrap() {
             PathState::UserSpacePath(ref path, Some(ref transform)) => {
-                Some(Box::new(path.as_azure().transformed_copy_to_builder(transform)))
+                Some(path.transformed_copy_to_builder(transform))
             },
             PathState::UserSpacePath(..) |
             PathState::UserSpacePathBuilder(..) |
@@ -973,8 +603,8 @@ impl<'a> CanvasData<'a> {
                         return;
                     },
                 };
-                let builder = Box::new(path.as_azure().transformed_copy_to_builder(&inverse));
-                Some(Path::Azure(builder.finish()))
+                let builder = path.transformed_copy_to_builder(&inverse);
+                Some(builder.finish())
             },
             PathState::UserSpacePathBuilder(..) | PathState::UserSpacePath(..) => None,
         };
@@ -993,7 +623,7 @@ impl<'a> CanvasData<'a> {
     }
 
     pub fn fill(&mut self) {
-        if is_zero_size_gradient(&self.state.fill_style) {
+        if self.state.fill_style.is_zero_size_gradient() {
             return; // Paint nothing if gradient size is zero.
         }
 
@@ -1006,7 +636,7 @@ impl<'a> CanvasData<'a> {
     }
 
     pub fn stroke(&mut self) {
-        if is_zero_size_gradient(&self.state.stroke_style) {
+        if self.state.stroke_style.is_zero_size_gradient() {
             return; // Paint nothing if gradient size is zero.
         }
 
@@ -1036,18 +666,18 @@ impl<'a> CanvasData<'a> {
             Some(PathState::UserSpacePath(ref path, ref transform)) => {
                 let target_transform = self.drawtarget.get_transform();
                 let path_transform = transform.as_ref().unwrap_or(&target_transform);
-                path.as_azure().contains_point(x, y, path_transform)
+                path.contains_point(x, y, path_transform)
             },
             Some(_) | None => false,
         };
         chan.send(result).unwrap();
     }
 
-    pub fn move_to(&mut self, point: &Point2D<AzFloat>) {
+    pub fn move_to(&mut self, point: &Point2D<f32>) {
         self.path_builder().move_to(point);
     }
 
-    pub fn line_to(&mut self, point: &Point2D<AzFloat>) {
+    pub fn line_to(&mut self, point: &Point2D<f32>) {
         self.path_builder().line_to(point);
     }
 
@@ -1069,14 +699,14 @@ impl<'a> CanvasData<'a> {
                 &PathState::UserSpacePathBuilder(ref builder, Some(ref transform)) => {
                     let path = builder.finish();
                     Some(PathState::DeviceSpacePathBuilder(
-                        Box::new(path.as_azure().transformed_copy_to_builder(transform)),
+                        path.transformed_copy_to_builder(transform),
                     ))
                 },
                 &PathState::UserSpacePath(ref path, Some(ref transform)) => Some(
-                    PathState::DeviceSpacePathBuilder(Box::new(path.as_azure().transformed_copy_to_builder(transform))),
+                    PathState::DeviceSpacePathBuilder(path.transformed_copy_to_builder(transform)),
                 ),
                 &PathState::UserSpacePath(ref path, None) => Some(PathState::UserSpacePathBuilder(
-                    Box::new(path.as_azure().copy_to_builder()),
+                    path.copy_to_builder(),
                     None,
                 )),
             }
@@ -1119,34 +749,34 @@ impl<'a> CanvasData<'a> {
         self.path_builder().rect(rect);
     }
 
-    pub fn quadratic_curve_to(&mut self, cp: &Point2D<AzFloat>, endpoint: &Point2D<AzFloat>) {
+    pub fn quadratic_curve_to(&mut self, cp: &Point2D<f32>, endpoint: &Point2D<f32>) {
         self.path_builder().quadratic_curve_to(cp, endpoint);
     }
 
     pub fn bezier_curve_to(
         &mut self,
-        cp1: &Point2D<AzFloat>,
-        cp2: &Point2D<AzFloat>,
-        endpoint: &Point2D<AzFloat>,
+        cp1: &Point2D<f32>,
+        cp2: &Point2D<f32>,
+        endpoint: &Point2D<f32>,
     ) {
         self.path_builder().bezier_curve_to(cp1, cp2, endpoint);
     }
 
     pub fn arc(
         &mut self,
-        center: &Point2D<AzFloat>,
-        radius: AzFloat,
-        start_angle: AzFloat,
-        end_angle: AzFloat,
+        center: &Point2D<f32>,
+        radius: f32,
+        start_angle: f32,
+        end_angle: f32,
         ccw: bool,
     ) {
         self.path_builder()
             .arc(center, radius, start_angle, end_angle, ccw);
     }
 
-    pub fn arc_to(&mut self, cp1: &Point2D<AzFloat>, cp2: &Point2D<AzFloat>, radius: AzFloat) {
+    pub fn arc_to(&mut self, cp1: &Point2D<f32>, cp2: &Point2D<f32>, radius: f32) {
         let cp0 = match self.path_builder().current_point() {
-            Some(p) => p.as_azure_point(),
+            Some(p) => p,
             None => return,
         };
         let cp1 = *cp1;
@@ -1209,12 +839,12 @@ impl<'a> CanvasData<'a> {
 
     pub fn ellipse(
         &mut self,
-        center: &Point2D<AzFloat>,
-        radius_x: AzFloat,
-        radius_y: AzFloat,
-        rotation_angle: AzFloat,
-        start_angle: AzFloat,
-        end_angle: AzFloat,
+        center: &Point2D<f32>,
+        radius_x: f32,
+        radius_y: f32,
+        rotation_angle: f32,
+        start_angle: f32,
+        end_angle: f32,
         ccw: bool,
     ) {
         self.path_builder().ellipse(
@@ -1229,15 +859,11 @@ impl<'a> CanvasData<'a> {
     }
 
     pub fn set_fill_style(&mut self, style: FillOrStrokeStyle) {
-        if let Some(pattern) = style.to_azure_pattern(&*self.drawtarget) {
-            self.state.fill_style = Pattern::Azure(pattern)
-        }
+        self.backend.set_fill_style(style, &mut self.state, &*self.drawtarget);
     }
 
     pub fn set_stroke_style(&mut self, style: FillOrStrokeStyle) {
-        if let Some(pattern) = style.to_azure_pattern(&*self.drawtarget) {
-            self.state.stroke_style = Pattern::Azure(pattern)
-        }
+        self.backend.set_stroke_style(style, &mut self.state, &*self.drawtarget);
     }
 
     pub fn set_line_width(&mut self, width: f32) {
@@ -1277,24 +903,12 @@ impl<'a> CanvasData<'a> {
     }
 
     pub fn set_global_composition(&mut self, op: CompositionOrBlending) {
-        self.state
-            .draw_options
-            .as_azure_mut()
-            .set_composition_op(op.to_azure_style());
-    }
-
-    pub fn create(size: Size2D<u64>) -> Box<GenericDrawTarget> {
-        // FIXME(nox): Why is the size made of i32 values?
-        Box::new(DrawTarget::new(
-            BackendType::Skia,
-            size.to_i32(),
-            azure_hl::SurfaceFormat::B8G8R8A8,
-        ))
+        self.backend.set_global_composition(op, &mut self.state);
     }
 
     pub fn recreate(&mut self, size: Size2D<u32>) {
-        self.drawtarget = CanvasData::create(Size2D::new(size.width as u64, size.height as u64));
-        self.state = CanvasPaintState::new(self.state.draw_options.as_azure().antialias, CanvasBackend::Azure);
+        self.drawtarget = self.backend.create_drawtarget(Size2D::new(size.width as u64, size.height as u64));
+        self.state = self.backend.recreate_paint_state(&self.state);
         self.saved_states.clear();
         // Webrender doesn't let images change size, so we clear the webrender image key.
         // TODO: there is an annying race condition here: the display list builder
@@ -1311,15 +925,15 @@ impl<'a> CanvasData<'a> {
 
     #[allow(unsafe_code)]
     pub fn send_pixels(&mut self, chan: IpcSender<IpcSharedMemory>) {
-        let data = IpcSharedMemory::from_bytes(unsafe {
-            self.drawtarget.snapshot().into_azure().get_data_surface().data()
-        });
+        let data = IpcSharedMemory::from_bytes(
+            &self.drawtarget.snapshot_data()
+        );
         chan.send(data).unwrap();
     }
 
     #[allow(unsafe_code)]
     pub fn send_data(&mut self, chan: IpcSender<CanvasImageData>) {
-        let size = self.drawtarget.get_size().into_azure();
+        let size = self.drawtarget.get_size();
 
         let descriptor = webrender_api::ImageDescriptor {
             size: webrender_api::DeviceIntSize::new(size.width, size.height),
@@ -1329,9 +943,9 @@ impl<'a> CanvasData<'a> {
             is_opaque: false,
             allow_mipmaps: false,
         };
-        let data = webrender_api::ImageData::Raw(Arc::new(unsafe {
-            self.drawtarget.snapshot().into_azure().get_data_surface().data().into()
-        }));
+        let data = webrender_api::ImageData::Raw(Arc::new(
+            self.drawtarget.snapshot_data().into()
+        ));
 
         let mut txn = webrender_api::Transaction::new();
 
@@ -1372,7 +986,6 @@ impl<'a> CanvasData<'a> {
                 &imagedata,
                 rect.size.to_i32(),
                 rect.size.width as i32 * 4,
-                SurfaceFormat::Azure(azure_hl::SurfaceFormat::B8G8R8A8),
             )
             .unwrap();
         self.drawtarget.copy_surface(
@@ -1394,13 +1007,13 @@ impl<'a> CanvasData<'a> {
         self.state.shadow_blur = value;
     }
 
-    pub fn set_shadow_color(&mut self, value: Color) {
-        self.state.shadow_color = value;
+    pub fn set_shadow_color(&mut self, value: RGBA) {
+        self.backend.set_shadow_color(value, &mut self.state);
     }
 
     // https://html.spec.whatwg.org/multipage/#when-shadows-are-drawn
     fn need_to_draw_shadow(&self) -> bool {
-        self.state.shadow_color.as_azure().a != 0.0f32 &&
+        self.backend.need_to_draw_shadow(&self.state.shadow_color) &&
             (self.state.shadow_offset_x != 0.0f64 ||
                 self.state.shadow_offset_y != 0.0f64 ||
                 self.state.shadow_blur != 0.0f64)
@@ -1431,16 +1044,16 @@ impl<'a> CanvasData<'a> {
         self.drawtarget.draw_surface_with_shadow(
             new_draw_target.snapshot(),
             &Point2D::new(
-                shadow_src_rect.origin.x as AzFloat,
-                shadow_src_rect.origin.y as AzFloat,
+                shadow_src_rect.origin.x as f32,
+                shadow_src_rect.origin.y as f32,
             ),
             &self.state.shadow_color,
             &Vector2D::new(
-                self.state.shadow_offset_x as AzFloat,
-                self.state.shadow_offset_y as AzFloat,
+                self.state.shadow_offset_x as f32,
+                self.state.shadow_offset_y as f32,
             ),
-            (self.state.shadow_blur / 2.0f64) as AzFloat,
-            CompositionOp::Azure(self.state.draw_options.as_azure().composition),
+            (self.state.shadow_blur / 2.0f64) as f32,
+            self.backend.get_composition_op(&self.state.draw_options),
         );
     }
 
@@ -1457,12 +1070,11 @@ impl<'a> CanvasData<'a> {
                     return vec![];
                 }
 
-        let data_surface = self.drawtarget.snapshot().into_azure().get_data_surface();
         pixels::rgba8_get_rect(
-            unsafe { data_surface.data() },
+            &self.drawtarget.snapshot_data(),
             canvas_size.to_u32(),
             read_rect.to_u32(),
-            )
+        )
             .into_owned()
     }
 }
@@ -1482,72 +1094,18 @@ impl<'a> Drop for CanvasData<'a> {
     }
 }
 
-enum CanvasBackend {
-    Azure,
-    Raqote,
-}
-
 #[derive(Clone)]
-struct CanvasPaintState<'a> {
-    draw_options: DrawOptions,
-    fill_style: Pattern,
-    stroke_style: Pattern,
-    stroke_opts: StrokeOptions<'a>,
+pub struct CanvasPaintState<'a> {
+    pub draw_options: DrawOptions,
+    pub fill_style: Pattern,
+    pub stroke_style: Pattern,
+    pub stroke_opts: StrokeOptions<'a>,
     /// The current 2D transform matrix.
-    transform: Transform2D<f32>,
-    shadow_offset_x: f64,
-    shadow_offset_y: f64,
-    shadow_blur: f64,
-    shadow_color: Color,
-}
-
-impl<'a> CanvasPaintState<'a> {
-    fn new(antialias: AntialiasMode, backend: CanvasBackend) -> CanvasPaintState<'a> {
-        match backend {
-            CanvasBackend::Azure => CanvasPaintState {
-                draw_options: DrawOptions::Azure(azure_hl::DrawOptions::new(1.0, azure_hl::CompositionOp::Over, antialias)),
-                fill_style: Pattern::Azure(azure_hl::Pattern::Color(ColorPattern::new(azure_hl::Color::black()))),
-                stroke_style: Pattern::Azure(azure_hl::Pattern::Color(ColorPattern::new(azure_hl::Color::black()))),
-                stroke_opts: StrokeOptions::Azure(azure_hl::StrokeOptions::new(
-                    1.0,
-                    JoinStyle::MiterOrBevel,
-                    CapStyle::Butt,
-                    10.0,
-                    &[],
-                    )),
-                    transform: Transform2D::identity(),
-                    shadow_offset_x: 0.0,
-                    shadow_offset_y: 0.0,
-                    shadow_blur: 0.0,
-                    shadow_color: Color::Azure(azure_hl::Color::transparent()),
-            },
-            CanvasBackend::Raqote => CanvasPaintState {
-                draw_options: DrawOptions::Raqote(()),
-                fill_style: Pattern::Raqote(()),
-                stroke_style: Pattern::Raqote(()),
-                stroke_opts: StrokeOptions::Raqote(()),
-                transform: Transform2D::identity(),
-                shadow_offset_x: 0.0,
-                shadow_offset_y: 0.0,
-                shadow_blur: 0.0,
-                shadow_color: Color::Raqote(()),
-            }
-        }
-    }
-}
-
-fn is_zero_size_gradient(pattern: &Pattern) -> bool {
-    match pattern {
-        Pattern::Azure(ref az_pattern) => {
-            if let azure_hl::Pattern::LinearGradient(ref gradient) = az_pattern {
-                if gradient.is_zero_size() {
-                    return true;
-                }
-            }
-            false
-        },
-        _ => unreachable!(),
-    }
+    pub transform: Transform2D<f32>,
+    pub shadow_offset_x: f64,
+    pub shadow_offset_y: f64,
+    pub shadow_blur: f64,
+    pub shadow_color: Color,
 }
 
 /// It writes an image to the destination target
@@ -1562,8 +1120,7 @@ fn write_image(
     image_size: Size2D<f64>,
     dest_rect: Rect<f64>,
     smoothing_enabled: bool,
-    composition_op: CompositionOp,
-    global_alpha: f32,
+    draw_options: &DrawOptions,
 ) {
     if image_data.is_empty() {
         return;
@@ -1586,17 +1143,14 @@ fn write_image(
             &image_data,
             image_size,
             image_size.width * 4,
-            SurfaceFormat::Azure(azure_hl::SurfaceFormat::B8G8R8A8),
-            )
+        )
         .unwrap();
-    let draw_surface_options = DrawSurfaceOptions::Azure(azure_hl::DrawSurfaceOptions::new(filter, true));
-    let draw_options =
-        DrawOptions::Azure(azure_hl::DrawOptions::new(global_alpha, composition_op.into_azure(), AntialiasMode::None));
+    
     draw_target.draw_surface(
         source_surface,
-        dest_rect.to_azure_style(),
-        image_rect.to_azure_style(),
-        draw_surface_options,
+        dest_rect,
+        image_rect,
+        filter,
         draw_options,
     );
 }
@@ -1644,191 +1198,6 @@ impl RectToi32 for Rect<f64> {
         Rect::new(
             Point2D::new(self.origin.x.ceil(), self.origin.y.ceil()),
             Size2D::new(self.size.width.ceil(), self.size.height.ceil()),
-        )
-    }
-}
-
-pub trait ToAzureStyle {
-    type Target;
-    fn to_azure_style(self) -> Self::Target;
-}
-
-impl ToAzureStyle for Rect<f64> {
-    type Target = Rect<AzFloat>;
-
-    fn to_azure_style(self) -> Rect<AzFloat> {
-        Rect::new(
-            Point2D::new(self.origin.x as AzFloat, self.origin.y as AzFloat),
-            Size2D::new(self.size.width as AzFloat, self.size.height as AzFloat),
-        )
-    }
-}
-
-impl ToAzureStyle for LineCapStyle {
-    type Target = CapStyle;
-
-    fn to_azure_style(self) -> CapStyle {
-        match self {
-            LineCapStyle::Butt => CapStyle::Butt,
-            LineCapStyle::Round => CapStyle::Round,
-            LineCapStyle::Square => CapStyle::Square,
-        }
-    }
-}
-
-impl ToAzureStyle for LineJoinStyle {
-    type Target = JoinStyle;
-
-    fn to_azure_style(self) -> JoinStyle {
-        match self {
-            LineJoinStyle::Round => JoinStyle::Round,
-            LineJoinStyle::Bevel => JoinStyle::Bevel,
-            LineJoinStyle::Miter => JoinStyle::Miter,
-        }
-    }
-}
-
-impl ToAzureStyle for CompositionStyle {
-    type Target = azure_hl::CompositionOp;
-
-    fn to_azure_style(self) -> azure_hl::CompositionOp {
-        match self {
-            CompositionStyle::SrcIn => azure_hl::CompositionOp::In,
-            CompositionStyle::SrcOut => azure_hl::CompositionOp::Out,
-            CompositionStyle::SrcOver => azure_hl::CompositionOp::Over,
-            CompositionStyle::SrcAtop => azure_hl::CompositionOp::Atop,
-            CompositionStyle::DestIn => azure_hl::CompositionOp::DestIn,
-            CompositionStyle::DestOut => azure_hl::CompositionOp::DestOut,
-            CompositionStyle::DestOver => azure_hl::CompositionOp::DestOver,
-            CompositionStyle::DestAtop => azure_hl::CompositionOp::DestAtop,
-            CompositionStyle::Copy => azure_hl::CompositionOp::Source,
-            CompositionStyle::Lighter => azure_hl::CompositionOp::Add,
-            CompositionStyle::Xor => azure_hl::CompositionOp::Xor,
-        }
-    }
-}
-
-impl ToAzureStyle for BlendingStyle {
-    type Target = azure_hl::CompositionOp;
-
-    fn to_azure_style(self) -> azure_hl::CompositionOp {
-        match self {
-            BlendingStyle::Multiply => azure_hl::CompositionOp::Multiply,
-            BlendingStyle::Screen => azure_hl::CompositionOp::Screen,
-            BlendingStyle::Overlay => azure_hl::CompositionOp::Overlay,
-            BlendingStyle::Darken => azure_hl::CompositionOp::Darken,
-            BlendingStyle::Lighten => azure_hl::CompositionOp::Lighten,
-            BlendingStyle::ColorDodge => azure_hl::CompositionOp::ColorDodge,
-            BlendingStyle::ColorBurn => azure_hl::CompositionOp::ColorBurn,
-            BlendingStyle::HardLight => azure_hl::CompositionOp::HardLight,
-            BlendingStyle::SoftLight => azure_hl::CompositionOp::SoftLight,
-            BlendingStyle::Difference => azure_hl::CompositionOp::Difference,
-            BlendingStyle::Exclusion => azure_hl::CompositionOp::Exclusion,
-            BlendingStyle::Hue => azure_hl::CompositionOp::Hue,
-            BlendingStyle::Saturation => azure_hl::CompositionOp::Saturation,
-            BlendingStyle::Color => azure_hl::CompositionOp::Color,
-            BlendingStyle::Luminosity => azure_hl::CompositionOp::Luminosity,
-        }
-    }
-}
-
-impl ToAzureStyle for CompositionOrBlending {
-    type Target = azure_hl::CompositionOp;
-
-    fn to_azure_style(self) -> azure_hl::CompositionOp {
-        match self {
-            CompositionOrBlending::Composition(op) => op.to_azure_style(),
-            CompositionOrBlending::Blending(op) => op.to_azure_style(),
-        }
-    }
-}
-
-pub trait ToAzurePattern {
-    fn to_azure_pattern(&self, drawtarget: &GenericDrawTarget) -> Option<azure_hl::Pattern>;
-}
-
-impl ToAzurePattern for FillOrStrokeStyle {
-    fn to_azure_pattern(&self, drawtarget: &GenericDrawTarget) -> Option<azure_hl::Pattern> {
-        Some(match *self {
-            FillOrStrokeStyle::Color(ref color) => {
-                azure_hl::Pattern::Color(ColorPattern::new(color.to_azure_style()))
-            },
-            FillOrStrokeStyle::LinearGradient(ref linear_gradient_style) => {
-                let gradient_stops: Vec<GradientStop> = linear_gradient_style
-                    .stops
-                    .iter()
-                    .map(|s| GradientStop::Azure(azure_hl::GradientStop {
-                        offset: s.offset as AzFloat,
-                        color: s.color.to_azure_style(),
-                    }))
-                    .collect();
-
-                azure_hl::Pattern::LinearGradient(LinearGradientPattern::new(
-                    &Point2D::new(
-                        linear_gradient_style.x0 as AzFloat,
-                        linear_gradient_style.y0 as AzFloat,
-                    ),
-                    &Point2D::new(
-                        linear_gradient_style.x1 as AzFloat,
-                        linear_gradient_style.y1 as AzFloat,
-                    ),
-                    drawtarget.create_gradient_stops(gradient_stops, ExtendMode::Azure(azure_hl::ExtendMode::Clamp)).into_azure(),
-                    &Transform2D::identity(),
-                ))
-            },
-            FillOrStrokeStyle::RadialGradient(ref radial_gradient_style) => {
-                let gradient_stops: Vec<GradientStop> = radial_gradient_style
-                    .stops
-                    .iter()
-                    .map(|s| GradientStop::Azure(azure_hl::GradientStop {
-                        offset: s.offset as AzFloat,
-                        color: s.color.to_azure_style(),
-                    }))
-                    .collect();
-
-                azure_hl::Pattern::RadialGradient(RadialGradientPattern::new(
-                    &Point2D::new(
-                        radial_gradient_style.x0 as AzFloat,
-                        radial_gradient_style.y0 as AzFloat,
-                    ),
-                    &Point2D::new(
-                        radial_gradient_style.x1 as AzFloat,
-                        radial_gradient_style.y1 as AzFloat,
-                    ),
-                    radial_gradient_style.r0 as AzFloat,
-                    radial_gradient_style.r1 as AzFloat,
-                    drawtarget.create_gradient_stops(gradient_stops, ExtendMode::Azure(azure_hl::ExtendMode::Clamp)).into_azure(),
-                    &Transform2D::identity(),
-                ))
-            },
-            FillOrStrokeStyle::Surface(ref surface_style) => {
-                let source_surface = drawtarget.create_source_surface_from_data(
-                    &surface_style.surface_data,
-                    // FIXME(nox): Why are those i32 values?
-                    surface_style.surface_size.to_i32(),
-                    surface_style.surface_size.width as i32 * 4,
-                    SurfaceFormat::Azure(azure_hl::SurfaceFormat::B8G8R8A8),
-                )?.into_azure();
-                azure_hl::Pattern::Surface(SurfacePattern::new(
-                    source_surface.azure_source_surface,
-                    surface_style.repeat_x,
-                    surface_style.repeat_y,
-                    &Transform2D::identity(),
-                ))
-            },
-        })
-    }
-}
-
-impl ToAzureStyle for RGBA {
-    type Target = azure_hl::Color;
-
-    fn to_azure_style(self) -> azure_hl::Color {
-        azure_hl::Color::rgba(
-            self.red_f32() as AzFloat,
-            self.green_f32() as AzFloat,
-            self.blue_f32() as AzFloat,
-            self.alpha_f32() as AzFloat,
         )
     }
 }
