+++ components/constellation/Cargo.toml
@@ -10,11 +10,15 @@ publish = false
 name = "constellation"
 path = "lib.rs"
 
+[features]
+azure_backend = ["canvas/azure_backend"]
+raqote_backend = ["canvas/raqote_backend"]
+
 [dependencies]
 background_hang_monitor = { path = "../background_hang_monitor"}
 backtrace = "0.3"
 bluetooth_traits = { path = "../bluetooth_traits" }
-canvas = {path = "../canvas"}
+canvas = {path = "../canvas", default-features = false}
 clipboard = "0.5"
 canvas_traits = {path = "../canvas_traits"}
 compositing = {path = "../compositing"}
